pyproject.toml
[tool.poetry]
name = "inotebook-backend"
version = "0.1.0"
description = ""
authors = ["ja-yy <jaypatel.learn@gmail.com>"]
readme = "README.md"
packages = [{include = "inotebook_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
pydantic = "^2.3.0"
motor = "^3.3.0"
pydantic-settings = "^2.0.3"
email-validator = "^2.0.0.post2"
pyyaml = "^6.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
